Add tests for AsyncValueCommand
